% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hs_available}
\alias{hs_available}
\title{Is Vectorscan or Hyperscan available?}
\usage{
hs_available()
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
Reports whether this installation was compiled with native
Vectorscan/Hyperscan support.
}
\examples{
hs_available()
}
