% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{hs_database_size}
\alias{hs_database_size}
\title{Get database size}
\usage{
hs_database_size(database)
}
\arguments{
\item{database}{A compiled \code{hs_database} object.}
}
\value{
Database size in bytes.
}
\description{
Get database size
}
\examples{
if (hs_available()) {
  db <- hs_database()
  hs_compile(db, "foo")
  hs_database_size(db)
}
}
