% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{hs_ext}
\alias{hs_ext}
\title{Create extended expression parameters}
\usage{
hs_ext(
  min_offset = NULL,
  max_offset = NULL,
  min_length = NULL,
  edit_distance = NULL,
  hamming_distance = NULL
)
}
\arguments{
\item{min_offset, max_offset, min_length}{Optional non-negative offsets.}

\item{edit_distance, hamming_distance}{Optional non-negative distances.}
}
\value{
An object that can be passed to \code{hs_compile()} through \code{ext}.
}
\description{
\code{hs_ext()} creates extended compile parameters for one expression.
}
\examples{
hs_ext(min_length = 3)
}
