% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{hs_info}
\alias{hs_info}
\title{Get database information}
\usage{
hs_info(database)
}
\arguments{
\item{database}{A compiled \code{hs_database} object.}
}
\value{
A single string reported by Vectorscan/Hyperscan.
}
\description{
Get database information
}
\examples{
if (hs_available()) {
  db <- hs_database()
  hs_compile(db, "foo")
  hs_info(db)
}
}
