% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{hs_scan_vector}
\alias{hs_scan_vector}
\title{Scan vectored data}
\usage{
hs_scan_vector(database, data, callback = NULL, context = NULL)
}
\arguments{
\item{database}{A compiled block-mode \code{hs_database}.}

\item{data}{A raw vector, character vector, or list of raw vectors/strings.}

\item{callback}{Optional function called as
\code{callback(id, from, to, flags, context)}. Return \code{TRUE} to stop scanning.}

\item{context}{Optional object passed to \code{callback}.}
}
\value{
A data frame of matches when \code{callback} is \code{NULL}; otherwise the
number of callback invocations, invisibly.
}
\description{
Scan vectored data
}
\examples{
if (hs_available()) {
  db <- hs_database(HS_MODE_VECTORED)
  hs_compile(db, "foobar")
  hs_scan_vector(db, c("foo", "bar"))
}
}
