% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{hs_stream_open}
\alias{hs_stream_open}
\title{Open a stream}
\usage{
hs_stream_open(database)
}
\arguments{
\item{database}{A compiled stream-mode \code{hs_database}.}
}
\value{
An \code{hs_stream} object.
}
\description{
Open a stream
}
\examples{
if (hs_available()) {
  db <- hs_database(HS_MODE_STREAM)
  hs_compile(db, "foo")
  stream <- hs_stream_open(db)
  hs_stream_scan(stream, "foo")
  hs_stream_close(stream)
}
}
