% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{hs_stream_size}
\alias{hs_stream_size}
\title{Size of a stream's state}
\usage{
hs_stream_size(database)
}
\arguments{
\item{database}{A compiled stream-mode \code{hs_database}.}
}
\value{
A single number of bytes.
}
\description{
The number of bytes of state each open stream of a stream-mode database
needs, a measure of the memory cost of keeping many streams open.
}
\examples{
if (hs_available()) {
  db <- hs_database(HS_MODE_STREAM)
  hs_compile(db, c("foo.*bar", "baz"))
  hs_stream_size(db)
}
}
