% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{hs_compile}
\alias{hs_compile}
\title{Compile expressions into a database}
\usage{
hs_compile(database, expressions, ids = NULL, flags = NULL, ext = NULL)
}
\arguments{
\item{database}{An \code{hs_database} object.}

\item{expressions}{Character vector of regular expressions.}

\item{ids}{Optional integer ids. Defaults to zero-based pattern positions.}

\item{flags}{Optional integer flags, recycled from length 1.}

\item{ext}{Optional \code{hs_ext()} object or list of \code{hs_ext()} objects.}
}
\value{
\code{database}, invisibly.
}
\description{
Compile expressions into a database
}
\examples{
if (hs_available()) {
  db <- hs_database()
  hs_compile(db, "foo")
  hs_scan(db, "foo")
}
}
