% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{hs_deserialize}
\alias{hs_deserialize}
\title{Deserialize a database}
\usage{
hs_deserialize(bytes)
}
\arguments{
\item{bytes}{A raw vector produced by \code{hs_serialize()}.}
}
\value{
A compiled \code{hs_database} object.
}
\description{
Deserialize a database
}
