% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{hs_load}
\alias{hs_load}
\title{Load a serialized database}
\usage{
hs_load(path)
}
\arguments{
\item{path}{File path to read.}
}
\value{
A compiled \code{hs_database} object.
}
\description{
Load a serialized database
}
