% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{hs_save}
\alias{hs_save}
\title{Save a serialized database}
\usage{
hs_save(database, path)
}
\arguments{
\item{database}{A compiled \code{hs_database}.}

\item{path}{File path to write.}
}
\value{
\code{path}, invisibly.
}
\description{
Save a serialized database
}
