% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{hs_serialize}
\alias{hs_serialize}
\title{Serialize a database}
\usage{
hs_serialize(database)
}
\arguments{
\item{database}{A compiled \code{hs_database}.}
}
\value{
A raw vector containing the serialized database.
}
\description{
Serialize a database
}
\examples{
if (hs_available()) {
  db <- hs_database()
  hs_compile(db, "foo")
  bytes <- hs_serialize(db)
  restored <- hs_deserialize(bytes)
  hs_scan(restored, "foo")
}
}
