% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{vectorscan-constants}
\alias{vectorscan-constants}
\alias{HS_FLAG_NONE}
\alias{HS_FLAG_CASELESS}
\alias{HS_FLAG_DOTALL}
\alias{HS_FLAG_MULTILINE}
\alias{HS_FLAG_SINGLEMATCH}
\alias{HS_FLAG_ALLOWEMPTY}
\alias{HS_FLAG_UTF8}
\alias{HS_FLAG_UCP}
\alias{HS_FLAG_PREFILTER}
\alias{HS_FLAG_SOM_LEFTMOST}
\alias{HS_FLAG_COMBINATION}
\alias{HS_FLAG_QUIET}
\alias{HS_MODE_BLOCK}
\alias{HS_MODE_STREAM}
\alias{HS_MODE_VECTORED}
\title{Vectorscan and Hyperscan constants}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
HS_FLAG_NONE

HS_FLAG_CASELESS

HS_FLAG_DOTALL

HS_FLAG_MULTILINE

HS_FLAG_SINGLEMATCH

HS_FLAG_ALLOWEMPTY

HS_FLAG_UTF8

HS_FLAG_UCP

HS_FLAG_PREFILTER

HS_FLAG_SOM_LEFTMOST

HS_FLAG_COMBINATION

HS_FLAG_QUIET

HS_MODE_BLOCK

HS_MODE_STREAM

HS_MODE_VECTORED
}
\description{
Integer constants used when compiling Vectorscan/Hyperscan databases.
}
\examples{
HS_FLAG_CASELESS
HS_MODE_BLOCK
}
\keyword{datasets}
