% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vectorscan-package}
\alias{vectorscan}
\alias{vectorscan-package}
\title{vectorscan}
\description{
R bindings to Vectorscan and Hyperscan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/vectorscan/}
  \item \url{https://github.com/pedrobtz/vectorscan}
  \item Report bugs at \url{https://github.com/pedrobtz/vectorscan/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Intel Corporation (Hyperscan, bundled in src/vendor/vectorscan) [copyright holder]
  \item VectorCamp PC (Vectorscan, bundled in src/vendor/vectorscan) [copyright holder]
  \item Arm Limited (Vectorscan, bundled in src/vendor/vectorscan) [copyright holder]
  \item University of Cambridge (PCRE2, bundled in src/vendor/pcre2) [copyright holder]
  \item Philip Hazel (PCRE2, bundled in src/vendor/pcre2) [copyright holder]
  \item Zoltan Herczeg (PCRE2 just-in-time compiler, bundled in src/vendor/pcre2) [copyright holder]
}

}
\keyword{internal}
