% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vectorscan-package}
\alias{vectorscan}
\alias{vectorscan-package}
\title{vectorscan}
\description{
R bindings to Vectorscan and Hyperscan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/vectorscan/}
  \item \url{https://github.com/pedrobtz/vectorscan}
  \item Report bugs at \url{https://github.com/pedrobtz/vectorscan/issues}
}

}
\author{
\strong{Maintainer}: pkg-vectorscan contributors \email{maintainer@example.com}

}
\keyword{internal}
